Move unmark to abuser file
